key generation for invite

+++ dds_web/api/schemas/user_schemas.py
@@ -14,6 +14,7 @@ import dds_web
 from dds_web import auth, db, utils
 from dds_web import errors as ddserr
 from dds_web.database import models
+from dds_web.security.project_user_keys import transfer_invite_private_key_to_user
 
 ####################################################################################################
 # SCHEMAS ################################################################################ SCHEMAS #
@@ -194,14 +195,33 @@ class NewUserSchema(marshmallow.Schema):
 
         db.session.add(new_user)
 
-        # TODO Use the TKEK here via the token.
-        # Idea is to handle it in auth.py together with project_user_keys.py
+        # Verify and transfer invite keys to the new user
+        temporary_key = dds_web.security.auth.verify_invite_key(token)
+        if temporary_key:
+            transfer_invite_private_key_to_user(invite, temporary_key, new_user)
+            for project_invite_key in invite.project_invite_keys:
+                project_user_key = models.ProjectUserKeys(
+                    project_id=project_invite_key.project_id,
+                    user_id=new_user.username,
+                    key=project_invite_key.key,
+                )
+                db.session.add(project_user_key)
+                db.session.delete(project_invite_key)
+
+            # TODO decrypt the user private key using the temp key,
+            #  derive a key from the password and encrypt the user private key with the derived key
+
+            flask.session.pop("invite_token", None)
+
+            # Delete old invite
+            db.session.delete(invite)
+
+            # Save and return
+            db.session.commit()
+
+            return new_user
+
         flask.session.pop("invite_token", None)
-
-        # Delete old invite
-        db.session.delete(invite)
-
-        # Save and return
+        db.session.delete(new_user)
         db.session.commit()
-
-        return new_user
+        raise ddserr.InviteError(message="Invite key verification has failed!")
